Removing all contracts when client or service was deleted;

+++ models/clients.php
@@ -1,4 +1,6 @@
 <?php
+    require_once('models/contracts.php');
+    
     class Clients {
         //make private
         public $id;
@@ -59,11 +61,15 @@
         }
 
         public static function remove($id) {
-            $db = Db::getInstance();
+            if (Contracts::removeByClient($id)) {
+                $db = Db::getInstance();
 
-            $stmt = $db->prepare("DELETE FROM users WHERE id = :id");
+                $stmt = $db->prepare("DELETE FROM users WHERE id = :id");
 
-            return $stmt->execute(array(':id' => $id));
+                return $stmt->execute(array(':id' => $id));
+            } else {
+                return false;
+            }
         }
 
         public static function getAll() {
